requirements_onnx_encoder.txt
Flask
Flask_Cors
gradio>=3.7.0
numpy==1.23.0
pyworld==0.2.5
scipy==1.10.0
SoundFile==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchcrepe
tqdm
loguru
scikit-maad
praat-parselmouth
onnx
onnxsim
onnxoptimizer
onnxruntime-gpu
librosa==0.9.1
tensorboard
tensorboardX
edge_tts
langdetect
pyyaml
pynvml
transformers
ffmpeg-python
faiss-cpu